RestakeTxDetails component test
